cli code and formatting cleanup
This replaces some of the more complex code in ara.utils with the
(hopefully) more obvious ara.fields.Field class, which also brings the
benefit of Jinja2 formatting to our CLI output.  This lets us utilize
jinja2 filters for formatting values for display, which avoids
code duplication and generally makes for cleaner display code.

This code also differentiates the output provided by the "list" and
"show" commands in several places such that the "show" command display
more information that is available in the corresponding "list"
command.